Override specialized ArrayDeque methods in Queue and Stack
There were three methods specialized in ArrayDeque to return some structure
of ArrayDeques instead of the underlying collection type. These were inherited
without overridng by Queue and Stack. This means the contract is broken: The type
says that e.g. sliding returns an Iterator[Queue] but in reality it returns
an Iterator[ArrayDeque].